- issue error messages with cmds without arguments are issued with arguments - added const declarations - marked unsued arguments with UNUSED

+++ nsthread/mutex.c
@@ -394,7 +394,7 @@ Ns_MutexList(Tcl_DString *dsPtr)
         Tcl_DStringStartSublist(dsPtr);
         Tcl_DStringAppendElement(dsPtr, mutexPtr->name);
         Tcl_DStringAppendElement(dsPtr, ""); /* unused? */
-        snprintf(buf, sizeof(buf),
+        snprintf(buf, (int)sizeof(buf),
                  " %" PRIuPTR " %lu %lu %" PRIu64 ".%06ld %" PRIu64 ".%06ld %" PRIu64 ".%06ld", 
                  mutexPtr->id, mutexPtr->nlock, mutexPtr->nbusy, 
                  (int64_t)mutexPtr->total_waiting_time.sec, mutexPtr->total_waiting_time.usec,
